pyproject.toml
[project]
name = "action-item-graph"
description = "Temporal knowledge graph pipeline for extracting and managing action items from call transcripts"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.11.0",
    "neo4j>=5.26.0",
    "openai>=1.50.0",
    "python-dotenv>=1.0.1",
    "tenacity>=9.0.0",
    "numpy>=1.26.0",
    "structlog>=24.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
    "pyright>=1.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "UP", "B", "SIM", "I"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"

[tool.pyright]
include = ["src"]
pythonVersion = "3.10"
typeCheckingMode = "basic"
